patient-service/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
passlib==1.7.4
bcrypt==4.0.1
pydantic[email]
python-dotenv
python-jose[cryptography]